// src/api/types.gen.ts
export type EventShortEntry = {
    id: string;
    name: string;
    description?: string | null;
    eventDateStart: string;
    eventDateEnd?: string | null;
    isCompleted: boolean;
    imageUri?: string | null;
};

export type EventShortEntryResponse = {
    data?: Array<EventShortEntry> | null;
    totalPages?: number;
    total?: number;
};

export type FileType = 'Jpg' | 'Jpeg' | 'Png' | 'Pdf' | 'Txt' | 'Doc' | 'Docx' | 'Xls' | 'Xlsx' | 'Link';

export type GetDebtEntry = {
    id?: string;
    amount?: number;
    isSent?: boolean;
    isComfirmed?: boolean;
    event?: EventShortEntry;
    lender?: UserShortEntry;
    debtor?: UserShortEntry;
};
// ...
export type GetUserEntry = {
    id?: string | null;
    name?: string | null;
    email?: string | null;
    role?: string | null;
    firstName?: string | null;
    middleName?: string | null;
    lastName?: string | null;
    phone?: string | null;
    imageUri?: string | null;
};
// ...
export type UserShortEntry = {
    id?: string | null;
    name?: string | null;
    imageUri?: string | null;
};